Fix drush cim/cex --skip-modules so that it also skips dependent configs.

+++ lib/Drush/Config/StorageWrapper.php
@@ -31,7 +31,11 @@ class StorageWrapper implements StorageInterface {
    * {@inheritdoc}
    */
   public function exists($name) {
-    return $this->storage->exists($name);
+    $exists = $this->storage->exists($name);
+    foreach ($this->filters as $filter) {
+      $exists = $filter->filterExists($name, $exists, $this->storage);
+    }
+    return $exists;
   }
 
   /**
@@ -41,7 +45,7 @@ class StorageWrapper implements StorageInterface {
     $data = $this->storage->read($name);
 
     foreach ($this->filters as $filter) {
-      $data = $filter->filterRead($name, $data);
+      $data = $filter->filterRead($name, $data, $this->storage);
     }
 
     return $data;
@@ -53,10 +57,15 @@ class StorageWrapper implements StorageInterface {
   public function readMultiple(array $names) {
     $dataList = $this->storage->readMultiple($names);
     $result = array();
-
+    // We also need to read the configs which are only on one of the two storages.
+    foreach($names as $name) {
+      if (!isset($dataList[$name])) {
+        $dataList[$name] = NULL;
+      }
+    }
     foreach ($dataList as $name => $data) {
       foreach ($this->filters as $filter) {
-        $data = $filter->filterRead($name, $data);
+        $data = $filter->filterRead($name, $data, $this->storage);
       }
       $result[$name] = $data;
     }
@@ -71,7 +80,10 @@ class StorageWrapper implements StorageInterface {
     foreach ($this->filters as $filter) {
       $data = $filter->filterWrite($name, $data, $this->storage);
     }
-
+    // filterWrite might return NULL, which means the config should be deleted.
+    if (!isset($data)) {
+      return $this->storage->delete($name);
+    }
     return $this->storage->write($name, $data);
   }
 
@@ -79,13 +91,23 @@ class StorageWrapper implements StorageInterface {
    * {@inheritdoc}
    */
   public function delete($name) {
-    return $this->storage->delete($name);
+    $doDelete = TRUE;
+    foreach ($this->filters as $filter) {
+      $doDelete = $filter->filterDelete($doDelete, $name, $this->storage);
+    }
+    if ($doDelete) {
+      return $this->storage->delete($name);
+    }
+    return TRUE;
   }
 
   /**
    * {@inheritdoc}
    */
   public function rename($name, $new_name) {
+    foreach ($this->filters as $filter) {
+      $new_name = $filter->filterRename($new_name, $name, $this->storage);
+    }
     return $this->storage->rename($name, $new_name);
   }
 
@@ -107,14 +129,24 @@ class StorageWrapper implements StorageInterface {
    * {@inheritdoc}
    */
   public function listAll($prefix = '') {
-    return $this->storage->listAll($prefix);
+    $list = $this->storage->listAll($prefix);
+    // Allow filters to add configs to the list.
+    foreach ($this->filters as $filter) {
+      $list = $filter->filterListAll($list, $this->storage, $prefix);
+    }
+    return $list;
   }
 
   /**
    * {@inheritdoc}
    */
   public function deleteAll($prefix = '') {
-    return $this->storage->deleteAll($prefix);
+    $list = $this->storage->listAll();
+    $result = TRUE;
+    foreach($list as $name) {
+      $result = $this->delete($name) ? $result : FALSE;
+    }
+    return $result;
   }
 
   /**
